Implement RowSet and RowRange.
These classes are required to specify the rows which should be read into
pyTorch's tensors.

These classes are only a thin wrapper around the C++ classes. It needs
to be done that way because eventually, those C++ classes need to be
passed to the underlying library. An alternative approach would be to
create everything in python and create the C++ implementations when
needed but it would be more complicated for no benefit.