Expose C++ RowSet and RowRange to Python.
These classes are required to specify the rows which should be read into
pyTorch's tensors.

These classes are exported from C++ to python directly. It needs to be
done that way because eventually, those C++ classes need to be passed to
the underlying library. An alternative approach would be to create
everything in python and create the C++ implementations when needed but
it would be more complicated for no benefit.