feat: version filters
We're adding version filters creation methods, so users can easily specify which versions they would like to include when reading the database.
We're wrapping in python only a subset of filter creation methods because the filter functionality that we need for this module is very limited. Other filters are expressed by other parts of the API (i.e. column filters are expressed by specifying columns as an argument in `read_rows` method) and user does not need direct access to them.